Update v2.0e (2026-02-24): New output module aof (ARM/Acorn Object Format, used by Acorn and 3DO); New CPU backend for Wozniak's Sweet16 pseudo CPU

+++ vobjdump.c
@@ -505,7 +505,7 @@ int main(int argc,char *argv[])
         }
         else
           fprintf(stderr,"Unable to allocate %lu bytes "
-                  "to buffer file \"%s\"!\n",vlen,argv[1]);
+                  "to buffer file \"%s\"!\n",(unsigned long)vlen,argv[1]);
       }
       fclose(f);
     }
